requirements.txt
asgiref==3.5.2
charset-normalizer==2.1.1
cloudinary==1.30.0
dj-database-url==1.0.0
Django==4.1.3
django-cloudinary-storage==0.3.0
django-environ==0.9.0
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.3.0
psycopg2==2.9.5
python-decouple==3.6
python-magic==0.4.27
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
whitenoise==6.2.0
